+++ src/lib/Components/Inbox/Conversations/Message.tsx
@@ -4,6 +4,9 @@ import * as Relay from "react-relay"
 
 import { BodyText, MetadataText, SmallHeadline } from "../Typography"
 
+import ImagePreview from "./Previews/ImagePreview"
+import PDFPreview from "./Previews/PDFPreview"
+
 import styled from "styled-components/native"
 import colors from "../../../../data/colors"
 
@@ -47,15 +50,36 @@ const ArtworkPreviewContainer = styled.View`marginBottom: 10;`
 
 const ImagePreviewContainer = styled.View`marginBottom: 10;`
 
+const PDFPreviewContainer = styled.View`marginBottom: 10;`
+
 interface Props extends RelayProps {
   senderName: string
   artworkPreview?: JSX.Element
-  imagePreview?: JSX.Element
 }
 
 export class Message extends React.Component<Props, any> {
+  renderAttachmentPreviews(attachments) {
+    return attachments.map(attachment => {
+      if (attachment.content_type.startsWith("image")) {
+        return (
+          <ImagePreviewContainer>
+            <ImagePreview imageAttachment={attachment} />
+          </ImagePreviewContainer>
+        )
+      }
+      if (attachment.content_type === "application/pdf") {
+        return (
+          <PDFPreviewContainer>
+            <PDFPreview pdfAttachment={attachment} />
+          </PDFPreviewContainer>
+        )
+      }
+    })
+  }
+
   render() {
     const date = moment(this.props.message.created_at).fromNow(true)
+
     return (
       <Container>
         <Avatar />
@@ -72,10 +96,9 @@ export class Message extends React.Component<Props, any> {
             <ArtworkPreviewContainer>
               {this.props.artworkPreview}
             </ArtworkPreviewContainer>}
-          {this.props.imagePreview &&
-            <ImagePreviewContainer>
-              {this.props.imagePreview}
-            </ImagePreviewContainer>}
+
+          {this.renderAttachmentPreviews(this.props.message.attachments)}
+
           <BodyText>
             {this.props.message.raw_text.split("\n\nAbout")[0]}
           </BodyText>
@@ -92,6 +115,11 @@ export default Relay.createContainer(Message, {
         raw_text
         created_at
         is_from_user
+        attachments {
+          content_type
+          download_url
+          file_name
+        }
       }
     `,
   },
@@ -102,5 +130,10 @@ interface RelayProps {
     raw_text: string | null
     created_at: string | null
     is_from_user: boolean
+    attachments: Array<{
+      content_type: string
+      download_url: string
+      file_name: string
+    }>
   }
 }