+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 -   [dev] Allows toggling the back button by pressing space - orta
 
 ###### Messaging
+
+-   Adds multiple attachment support - matt
+-   Adds PDF Preview for attachments - matt
 -   Fixes pagination for Conversations - luc
 -   Adds pull to refresh support to Inbox view - luc
 -   Adds an `ImagePreview` to each Conversation, Relay-ified messages - matt